refactor: simplify validation handling in CreditRequest interface
