+++ src/pages/board/outlets/financialReporting/Requirements/index.tsx
@@ -7,82 +7,34 @@ import {
 import { Icon } from "@inubekit/icon";
 import { Flag } from "@inubekit/flag";
 import { Stack } from "@inubekit/stack";
-import { Tag } from "@inubekit/tag";
 
 import { Fieldset } from "@components/data/Fieldset";
 import { TableBoard } from "@components/data/TableBoard";
 import { IAction, IEntries, ITitle } from "@components/data/TableBoard/types";
-import { titlesRequirements } from "@config/pages/board/oulet/financialReporting/configRequirements";
 import { getById } from "@mocks/utils/dataMock.service";
 import { CreditRequest } from "@services/types";
 
-import { actionsMobile, dataButton, infoItems } from "./config";
+import {
+  actionsMobile,
+  dataButton,
+  infoItems,
+  maperDataRequirements,
+  maperEntries,
+} from "./config";
 import { SeeDetailsModal } from "./SeeDetailsModal";
 import { AprovalsModal } from "./AprovalsModal";
 import { StyledMessageContainer } from "../styles";
 
-interface IData {
+interface IRequirementsData {
   id: string;
   titlesRequirements: ITitle[];
   entriesRequirements: IEntries[];
-  actionsRequirements: IAction[];
   actionsMovile: IAction[];
 }
 
 export interface IRequirementsProps {
-  data: IData[];
   isMobile: boolean;
   id: string;
-}
-
-const generateTag = (value: string): JSX.Element => {
-  if (value === "Y") {
-    return <Tag label="Cumple" appearance="success" weight="strong" />;
-  } else if (value === "N") {
-    return <Tag label="No Cumple" appearance="danger" weight="strong" />;
-  } else {
-    return <Tag label="Sin Evaluar" appearance="warning" weight="strong" />;
-  }
-};
-
-const maperEntries = (data: CreditRequest): IEntries[][] => {
-  const result: IEntries[][] = [];
-
-  const systemValidations: IEntries[] = Object.entries(
-    data.system_validations
-  ).map(([key, value], index) => ({
-    id: `sistema-${index + 1}`,
-    "Validaciones del sistema": key,
-    tag: generateTag(value),
-  }));
-
-  const documentaryRequirements: IEntries[] = Object.entries(
-    data.documentary_requirements
-  ).map(([key, value], index) => ({
-    id: `documento-${index + 1}`,
-    "Requisitos documentales": key,
-    tag: generateTag(value),
-  }));
-
-  const humanValidations: IEntries[] = Object.entries(
-    data.human_validations
-  ).map(([key, value], index) => ({
-    id: `humano-${index + 1}`,
-    "Validaciones humanas": key,
-    tag: generateTag(value),
-  }));
-
-  result.push(systemValidations, documentaryRequirements, humanValidations);
-  console.log("result", result);
-
-  return result;
-};
-
-interface dataRequirements {
-  id: string;
-  titlesRequirements: ITitle[];
-  entriesRequirements: IEntries[];
-  actionsMovile: IAction[];
 }
 
 export const Requirements = (props: IRequirementsProps) => {
@@ -95,7 +47,7 @@ export const Requirements = (props: IRequirementsProps) => {
   const [isApproved, setIsApproved] = useState(false);
   const [showFlagMessage, setShowFlagMessage] = useState(false);
 
-  const [dataRequirements, setDataRequirements] = useState<dataRequirements[]>(
+  const [dataRequirements, setDataRequirements] = useState<IRequirementsData[]>(
     []
   );
 
@@ -110,28 +62,9 @@ export const Requirements = (props: IRequirementsProps) => {
 
         if (!(requirements instanceof Error)) {
           const processedEntries = maperEntries(requirements);
-          console.log("prueba", typeof processedEntries);
+          const processedRequirements = maperDataRequirements(processedEntries);
 
-          setDataRequirements([
-            {
-              id: "tabla1",
-              titlesRequirements: titlesRequirements[0],
-              entriesRequirements: processedEntries[0],
-              actionsMovile: actionsMobile,
-            },
-            {
-              id: "tabla2",
-              titlesRequirements: titlesRequirements[1],
-              entriesRequirements: processedEntries[1],
-              actionsMovile: actionsMobile,
-            },
-            {
-              id: "tabla3",
-              titlesRequirements: titlesRequirements[2],
-              entriesRequirements: processedEntries[2],
-              actionsMovile: actionsMobile,
-            },
-          ]);
+          setDataRequirements(processedRequirements);
         }
       } catch (error) {
         console.error(error);
@@ -197,8 +130,6 @@ export const Requirements = (props: IRequirementsProps) => {
     { id: "agregar", content: renderAddIcon },
     { id: "aprobar", content: renderCheckIcon },
   ];
-
-  console.log("dataRequirements", dataRequirements);
 
   return (
     <>
